gwpy.testing: moved mock import to compat module
