+++ gwpy/cli/tests/base.py
@@ -34,7 +34,7 @@ from ...frequencyseries import FrequencySeries
 from ...timeseries import TimeSeries
 from ...plot import Plot
 from ...testing import (utils, mocks)
-from ...testing.mocks import mock
+from ...testing.compat import mock
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
