+++ gwpy/cli/tests/test_qtransform.py
@@ -22,8 +22,8 @@
 import os
 
 from ... import cli
+from ...testing.compat import mock
 from ...testing.utils import skip_missing_dependency
-from ...testing.mocks import mock
 from .base import (update_namespace, mock_nds2_connection)
 from .test_spectrogram import TestCliSpectrogram as _TestCliSpectrogram
