+++ gwpy/plot/tests/test_tex.py
@@ -22,7 +22,7 @@
 import pytest
 
 from .. import tex as plot_tex
-from ...testing.mocks import mock
+from ...testing.compat import mock
 
 
 def which_patcher(error=False):
