+++ gwpy/segments/tests/test_flag.py
@@ -36,7 +36,7 @@ from ...plot import SegmentAxes
 from ...segments import (Segment, SegmentList,
                          DataQualityFlag, DataQualityDict)
 from ...testing import (mocks, utils)
-from ...testing.mocks import mock
+from ...testing.compat import mock
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
