+++ gwpy/table/tests/test_table.py
@@ -42,7 +42,7 @@ from ...frequencyseries import FrequencySeries
 from ...io import ligolw as io_ligolw
 from ...segments import (Segment, SegmentList)
 from ...testing import utils
-from ...testing.mocks import mock
+from ...testing.compat import mock
 from ...time import LIGOTimeGPS
 from ...timeseries import (TimeSeries, TimeSeriesDict)
 from .. import (Table, EventTable, filters)
