+++ gwpy/timeseries/tests/test_timeseries.py
@@ -41,7 +41,7 @@ from ...segments import Segment
 from ...signal import filter_design
 from ...spectrogram import Spectrogram
 from ...testing import (mocks, utils)
-from ...testing.mocks import mock
+from ...testing.compat import mock
 from ...time import LIGOTimeGPS
 from ...utils.misc import null_context
 from .. import (TimeSeries, TimeSeriesDict, TimeSeriesList, StateTimeSeries)
